requirements.txt
Django
psycopg2-binary
requests
python-dotenv
Pillow
django-widget-tweaks
beautifulsoup4
playwright
openpyxl
social-auth-app-django
whitenoise
gunicorn